refactor: rest형식으로 리펙토링
